[MISC][WL] create useValidationConfig hook - hook is for formValidationConfig only - update useValidationSchema to handle schema generation only
