+++ src/components/fields/checkbox-group/checkbox-group.tsx
@@ -3,7 +3,7 @@ import without from "lodash/without";
 import React, { useEffect, useState } from "react";
 import * as Yup from "yup";
 import { TestHelper } from "../../../utils";
-import { useValidationSchema } from "../../../utils/hooks";
+import { useValidationConfig } from "../../../utils/hooks";
 import { IGenericFieldProps } from "../../frontend-engine";
 import { ERROR_MESSAGES } from "../../shared";
 import { CheckboxContainer, Label, StyledCheckbox } from "./checkbox-group.styles";
@@ -22,7 +22,7 @@ export const CheckboxGroup = (props: IGenericFieldProps<ICheckboxGroupSchema>) =
 	} = props;
 
 	const [stateValue, setStateValue] = useState<string[]>(value || []);
-	const { setFieldValidationConfig } = useValidationSchema();
+	const { setFieldValidationConfig } = useValidationConfig();
 
 	// =============================================================================
 	// EFFECTS
