+++ src/components/fields/date-input/date-input.tsx
@@ -3,7 +3,7 @@ import { Form } from "@lifesg/react-design-system/form";
 import { useEffect, useState } from "react";
 import * as Yup from "yup";
 import { DateTimeHelper, TestHelper } from "../../../utils";
-import { useValidationSchema } from "../../../utils/hooks";
+import { useValidationConfig } from "../../../utils/hooks";
 import { IGenericFieldProps } from "../../frontend-engine/types";
 import { ERROR_MESSAGES } from "../../shared";
 import { IDateInputSchema } from "./types";
@@ -21,7 +21,7 @@ export const DateInput = (props: IGenericFieldProps<IDateInputSchema>) => {
 		...otherProps
 	} = props;
 	const [stateValue, setStateValue] = useState<string>(value || ""); // adheres to dateFormat
-	const { setFieldValidationConfig } = useValidationSchema();
+	const { setFieldValidationConfig } = useValidationConfig();
 
 	// =============================================================================
 	// EFFECTS
