+++ src/components/fields/submit-button/submit-button.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@lifesg/react-design-system/button";
 import { useEffect } from "react";
 import * as Yup from "yup";
-import { useValidationSchema } from "../../../utils/hooks";
+import { useValidationConfig } from "../../../utils/hooks";
 import { IGenericFieldProps } from "../../frontend-engine";
 import { ISubmitButtonSchema } from "./types";
 
@@ -14,7 +14,7 @@ export const SubmitButton = (props: IGenericFieldProps<ISubmitButtonSchema>) => 
 		id,
 		...otherProps
 	} = props;
-	const { setFieldValidationConfig } = useValidationSchema();
+	const { setFieldValidationConfig } = useValidationConfig();
 
 	// =============================================================================
 	// EFFECTS
