+++ src/utils/hooks/index.ts
@@ -1,1 +1,2 @@
+export * from "./use-validation-config";
 export * from "./use-validation-schema";
